backend/fastapi-backend-database/requirements copy.txt
fastapi
uvicorn
pymongo[srv]==3.12
python-dotenv